requirements.txt
asgiref==3.5.2
Django==4.1.2
django-crispy-forms==1.14.0
Pillow==9.2.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
tzdata==2022.5
